bump: spreadsheet logic.
