+++ src/lib/layout/shell.svelte
@@ -230,6 +230,15 @@
         }
     }
 
+    :global(main:has(.databases-spreadsheet)) {
+        @media (min-width: 1024px) {
+            .main-content {
+                height: auto;
+                padding-left: 210px;
+            }
+        }
+    }
+
     .overlay {
         position: fixed;
         width: 100vw;
