+++ src/routes/(console)/project-[region]-[project]/+layout.svelte
@@ -126,7 +126,7 @@
 <style>
     .layout-level-progress-bars {
         gap: 1rem;
-        z-index: 1;
+        z-index: 2;
         display: flex;
         flex-direction: column;
